+++ src/components/Button/styles.ts
@@ -1,6 +1,7 @@
 export const variants = {
   primary:
     "border-purple-400 hover:bg-purple-500 hover:text-white text-purple-500",
+  active: "border-purple-500 bg-purple-500 text-white",
   danger: "border-red-500 bg-red-400 text-zinc-50 hover:opacity-80",
   ghost: "border-transparent text-zinc-400 hover:text-white",
 };